requirements.txt
# Orchestration
apache-airflow==2.10.4

# LLM
anthropic>=0.40.0

# Google Cloud
google-cloud-bigquery>=3.25.0
google-cloud-storage>=2.18.0
db-dtypes>=1.3.0

# YouTube
youtube-transcript-api>=0.6.3
google-api-python-client>=2.150.0

# Database & Vector Store
psycopg2-binary>=2.9.9
pgvector>=0.3.5

# API
fastapi>=0.115.0
uvicorn>=0.32.0

# Chat UI
streamlit>=1.40.0

# Embeddings (local, free)
sentence-transformers>=3.0.0

# Data processing
pandas>=2.2.0
tiktoken>=0.8.0

# Testing
pytest>=8.3.0
